+++ src/main/webapp/chapter3/index-mvvm.zul
@@ -8,7 +8,7 @@
 			<include src="/chapter2/sidebar.zul"/>
 		</west>
 		<center id="mainContent" autoscroll="true" border="none">
-			<include src="/chapter5/profile-mvvm.zul"/>
+			<include src="/chapter3/profile-mvvm.zul"/>
 		</center>
 		<south height="50px" border="none">
 			<include src="/chapter1/footer.zul"/>
