+++ src/main/webapp/chapter3/profile-mvc.zul
@@ -1,5 +1,5 @@
 <?link rel="stylesheet" type="text/css" href="/style.css"?>
-<window apply="org.zkoss.essentials.chapter5.mvc.ProfileViewController" 
+<window apply="org.zkoss.essentials.chapter3.mvc.ProfileViewController" 
 	border="normal" hflex="1" vflex="1" contentStyle="overflow:auto">
 	<caption src="/imgs/profile.png" sclass="fn-caption" label="Profile (MVC)"/>
 	<vlayout>
