+++ src/main/webapp/chapter3/profile-mvvm.zul
@@ -1,13 +1,10 @@
-<!--
-	An example to use property binding instead of form binding. 
- -->
 <?link rel="stylesheet" type="text/css" href="/style.css"?>
 <window apply="org.zkoss.bind.BindComposer"
-	viewModel="@id('vm') @init('org.zkoss.essentials.chapter5.mvvm.ProfileViewModel')" 
+	viewModel="@id('vm') @init('org.zkoss.essentials.chapter3.mvvm.ProfileViewModel')" 
 	border="normal" hflex="1" vflex="1" contentStyle="overflow:auto">
 	<caption src="/imgs/profile.png" sclass="fn-caption" label="Profile (MVVM)"/>
 	<vlayout>
-		<grid width="500px">
+		<grid width="500px" form="@id('fx') @load(vm.currentUser) @save(vm.currentUser, before='save')">
 			<columns>
 				<column align="right" hflex="min"/>
 				<column/>
@@ -15,24 +12,34 @@
 			<rows>
 				<row>
 					<cell sclass="row-title">Account :</cell>
-					<cell><label value="@load(vm.currentUser.account)"/></cell>
+					<cell><label value="@load(fx.account)"/></cell>
 				</row>
 				<row>
 					<cell sclass="row-title">Full Name :</cell>
-					<cell><textbox value="@bind(vm.currentUser.fullName)" constraint="no empty: Plean enter your full name" width="200px"/></cell>
+					<cell>
+						<textbox value="@bind(fx.fullName)" width="200px"
+						constraint="no empty: Plean enter your full name" />
+					</cell>
 				</row>
 				<row>
 					<cell sclass="row-title">Email :</cell>
-					<cell><textbox value="@bind(vm.currentUser.email)" constraint="/.+@.+\.[a-z]+/: Please enter an e-mail address" width="200px"/></cell>
+					<cell>
+						<textbox value="@bind(fx.email)" width="200px"
+						constraint="/.+@.+\.[a-z]+/: Please enter an e-mail address" />
+					</cell>
 				</row>
 				<row>
 					<cell sclass="row-title">Birthday :</cell>
-					<cell><datebox value="@bind(vm.currentUser.birthday)" constraint="no future" width="200px"/></cell>
+					<cell>
+						<datebox value="@bind(fx.birthday)" width="200px"
+						constraint="no future" />
+					</cell>
 				</row>
 				<row>
 					<cell sclass="row-title">Country :</cell>
 					<cell>
-						<listbox model="@load(vm.countryList)" selectedItem="@bind(vm.currentUser.country)" mold="select" width="200px">
+						<listbox model="@load(vm.countryList)" mold="select" width="200px"
+						 selectedItem="@bind(fx.country)">
 							<template name="model">
 								<listitem label="@load(each)" />
 							</template>
@@ -41,7 +48,7 @@
 				</row>
 				<row>
 					<cell sclass="row-title">Bio :</cell>
-					<cell><textbox value="@bind(vm.currentUser.bio)" multiline="true" hflex="1" height="200px" /></cell>
+					<cell><textbox value="@bind(fx.bio)" multiline="true" hflex="1" height="200px" /></cell>
 				</row>
 			</rows>
 		</grid>
